requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
idna==3.10
openpyxl==3.1.2
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
reportlab==4.2.2
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
